requirements.txt
requests
flask
python-dotenv
flask_cors